Divided codes into separate files
With this commit
 - class Hero will be saved in Hero.js file
 - maps will be saved in maps.json file
