+++ MazeEscape/index.js
@@ -1,68 +1,13 @@
 const readline = require('readline');
 const esc = require('ansi-escapes');
+const maps = require('./maps.json').maps;
+const Hero = require('./Hero');
 
-class Hero {
+let round = 0;
 
-    constructor() {
-        this.x = 0;
-        this.y = 0;
-    }
+let map = maps[round];
 
-    moveUp() {
-        if (this.canGo(this.x, this.y - 1)) {
-            this.leave();
-            this.goto(this.x, this.y - 1);
-        }
-    }
-
-    moveDown()  {
-        if (this.canGo(this.x, this.y + 1)) {
-            this.leave();
-            this.goto(this.x, this.y + 1);
-        }
-    }
-
-    moveLeft() {
-        if (this.canGo(this.x - 1, this.y)) {
-            this.leave();
-            this.goto(this.x - 1, this.y);
-        }
-    }
-
-    moveRight() {
-        if (this.canGo(this.x + 1, this.y)) {
-            this.leave();
-            this.goto(this.x + 1, this.y);
-        }
-    }
-
-    goto(x, y) {
-        if (this.canGo(x, y)) {
-            this.x = x;
-            this.y = y;
-            map[y][x] = 'X';
-        }
-    }
-
-    leave()  {
-        map[this.y][this.x] = " ";
-    }
-
-    canGo(x, y) {
-        return map[y] && map[y][x] == " ";
-    }
-
-}
-
-let map = [
-    ["X", " ", " "],
-    [" ", " ", " "],
-    [" ", " ", " "],
-    [" ", " ", " "]
-]
-
-
-let hero = new Hero();
+let hero = new Hero(map);
 
 readline.emitKeypressEvents(process.stdin);
 process.stdin.setRawMode(true);
@@ -95,7 +40,7 @@ function draw() {
     let res = '';
     for (let row of map) {
         for (let col of row) {
-            res += col;
+            res += col + " ";
         }
         res += '\n';
     }
