+++ MazeEscape/maps.json
@@ -1,0 +1,9 @@
+{
+    "maps": [
+        [
+            ["♟", " ", " ", " "],
+            ["▉", "▉" ," ", " "],
+            [" ", "▉", "▉", "▉"]
+        ]
+    ]
+}
